[refactor] Deprecated `add_project_header_command` (6m)
