+++ lib/houston/boot/extensions.rb
@@ -5,15 +5,14 @@ require "houston/boot/extensions/oauth"
 require "houston/boot/extensions/features"
 require "houston/boot/extensions/serializers"
 require "houston/boot/extensions/view"
-require "houston/boot/extensions/deprecated_methods"
+require "houston/boot/extensions/deprecated"
 require "houston/boot/serializers/active_record_serializer"
 require "houston/boot/serializers/readonly_hash_serializer"
 
 
 module Houston
   module Extensions
-    include Houston::Extensions::DeprecatedMethods
-
+    include Houston::Extensions::Deprecated
 
     def events
       return @events if defined?(@events)
@@ -51,7 +50,6 @@ module Houston
     end
 
 
-
     def register_events(&block)
       events.register(&block)
     end
@@ -60,87 +58,34 @@ module Houston
       serializers.add(serializer)
     end
 
-
-
-
-
-    def add_project_header_command(slug, &block)
-      dsl = ProjectBannerFeatureDsl.new(ProjectBannerFeature.new)
-      dsl.instance_eval(&block)
-      feature = dsl.feature
-      feature.slug = slug
-
-      @project_header_commands[slug] = feature
-    end
-
-    def project_header_commands
-      @project_header_commands.values
-    end
-
-
-
-  private
-
-    class ProjectBannerFeature
-      attr_accessor :partial
-      attr_accessor :ability_block
-      attr_accessor :slug
-
-      def permitted?(ability, project)
-        return true if ability_block.nil?
-        ability_block.call ability, project
-      end
-    end
-
-    class ProjectBannerFeatureDsl
-      attr_reader :feature
-
-      def initialize(feature)
-        @feature = feature
-      end
-
-      def partial(value)
-        feature.partial = value
-      end
-
-      def ability(&block)
-        feature.ability_block = block
-      end
-    end
-
   end
 
 
 
-  @project_header_commands = {}
-  extend Houston::Extensions
+  extend Extensions
+
+  view["projects"].has :Table
+  view["users"].has :Table
+  view["edit_project"].has :Form
+  view["edit_user"].has :Form
+
+  project_features
+    .add(:settings) { |project| Houston::Application.routes.url_helpers.edit_project_path(project) }
+    .ability { |project| can?(:update, project) }
+
+  register_events {{
+    "daemon:{type}:start"   => desc("Daemon {type} has started"),
+    "daemon:{type}:restart" => desc("Daemon {type} has restarted"),
+    "daemon:{type}:stop"    => desc("Daemon {type} has stopped"),
+
+    "hooks:{type}"          => params("params").desc("/hooks/{type} was invoked"),
+    "hooks:project:{type}"  => params("project", "params").desc("/hooks/project/:slug/{type} was invoked"),
+
+    "authorization:grant"   => params("authorization").desc("Authorization was granted"),
+    "authorization:revoke"  => params("authorization").desc("Authorization was revoked")
+  }}
+
+  serializers << Houston::ActiveRecordSerializer.new
+  serializers << Houston::ReadonlyHashSerializer.new
+
 end
-
-
-
-Houston.view["projects"].has :Table
-Houston.view["users"].has :Table
-Houston.view["edit_project"].has :Form
-Houston.view["edit_user"].has :Form
-
-Houston.project_features
-  .add(:settings) { |project| Houston::Application.routes.url_helpers.edit_project_path(project) }
-  .ability { |project| can?(:update, project) }
-
-Houston.register_events {{
-
-  "daemon:{type}:start"             => desc("Daemon {type} has started"),
-  "daemon:{type}:restart"           => desc("Daemon {type} has restarted"),
-  "daemon:{type}:stop"              => desc("Daemon {type} has stopped"),
-
-  "hooks:{type}"                    => params("params").desc("/hooks/{type} was invoked"),
-  "hooks:project:{type}"            => params("project", "params").desc("/hooks/project/:slug/{type} was invoked"),
-
-  "authorization:grant"             => params("authorization").desc("Authorization was granted"),
-  "authorization:revoke"            => params("authorization").desc("Authorization was revoked")
-
-}}
-
-Houston.serializers << Houston::ActiveRecordSerializer.new
-
-Houston.serializers << Houston::ReadonlyHashSerializer.new
